+++ lib/barbeque/message_handler/notification.rb
@@ -9,7 +9,7 @@ module Barbeque
         @message = message
         @message_queue = message_queue
 
-        subscription = SNSSubscription.find_by!(topic_arn: @message.topic_arn, job_queue_id: @message_queue.job_queue.id)
+        subscription = SnsSubscription.find_by!(topic_arn: @message.topic_arn, job_queue_id: @message_queue.job_queue.id)
         @message.set_params_from_subscription(subscription)
       end
     end